// packages/joystick/src/types/db.types.ts
export enum Collections {
  Authorigins = "_authOrigins",
  Externalauths = "_externalAuths",
  Mfas = "_mfas",
  Otps = "_otps",
  Superusers = "_superusers",
  Actions = "actions",
  ActionLogs = "action_logs",
  Devices = "devices",
  Models = "models",
  Permissions = "permissions",
  Rules = "rules",
  Run = "run",
  Templates = "templates",
  Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string;
export type RecordIdString = string;
export type HTMLString = string;

// System fields
export type BaseSystemFields<T = never> = {
  id: RecordIdString;
  collectionId: string;
  collectionName: Collections;
  expand?: T;
};

export type AuthSystemFields<T = never> = {
  email: string;
  emailVisibility: boolean;
  username: string;
  verified: boolean;
} & BaseSystemFields<T>;
// ...
export type ActionsRecord = {
  created?: IsoDateString;
  id: string;
  name: string;
  updated?: IsoDateString;
};

export enum DevicesStatusOptions {
  "off" = "off",
  "on" = "on",
  "waiting" = "waiting",
}
export type DevicesRecord<
  Tautomation = unknown,
  Tconfiguration = unknown,
  Tinformation = unknown
> = {
  allow?: RecordIdString[];
  auto?: boolean;
  automation?: null | Tautomation;
  configuration?: null | Tconfiguration;
  created?: IsoDateString;
  description?: HTMLString;
  device: RecordIdString;
  hide?: boolean;
  id: string;
  information: null | Tinformation;
  mode: string;
  name?: string;
  status?: DevicesStatusOptions;
  updated?: IsoDateString;
};
// ...
export type UsersRecord = {
  avatar?: string;
  created?: IsoDateString;
  email: string;
  emailVisibility?: boolean;
  id: string;
  name?: string;
  password: string;
  tokenKey: string;
  updated?: IsoDateString;
  verified?: boolean;
};
// ...
export type ActionsResponse<Texpand = unknown> = Required<ActionsRecord> &
  BaseSystemFields<Texpand>;
export type DevicesResponse<
  Tautomation = unknown,
  Tconfiguration = unknown,
  Tinformation = unknown,
  Texpand = unknown
> = Required<DevicesRecord<Tautomation, Tconfiguration, Tinformation>> &
  BaseSystemFields<Texpand>;
// ...
export type UsersResponse<Texpand = unknown> = Required<UsersRecord> &
  AuthSystemFields<Texpand>;